RainGarden and Bioswale drainage area
Migrations to add drainage_area field to RainGarden and Bioswale, and
update the adjuncts_timestamp to force reloads of instances with prior
RainGardens and Bioswales, which forces default permission updating.

Add unit settings.

Add templates and remove skip directive, to get RainGardens and Bioswales
to display in the resource and resource add sidebars.

Update tests.

--

Connects to #2688