Update stormwater benefit calculation to include drainage area
Note: you might still see incorrect aggregate results in your browser due to
OpenTreeMap/otm-core#2707. Going in with the shell and kicking universal_revs a
couple times seems to resolve it.

Added the new calculation to stormwater/benefits.py. Since drainage_area is not
in PolygonalMapFeature, had to change the interface to the internal
_benefits_for_feature_qs and pass in a qs of the original map feature class.

The new calculation also calls for a default constant of 0.85.

Updated tests accordingly.

--

Connected to #2689