Add meta species lookup table for Tampa migration
I added a static lookup table in the Tampa migration code to handle
species that do not have an i-Tree code set in the species fixture. I
opted to handle these 7 edge case species with a literal lookup
rather than make the non-trivial changes to the migrator to properly
handle the one-to-many relationship between species rows and resource
rows. The migrator is reaching the end of its life and the cost of the
changes does not justify the relatively small gain in general
correctness.