PR fixup: fix date initialization
So it turns out that date pickers are initialized automatically on load --
you don't have to call `datepicker()`. So the code that called `datepicker` on
form initialization for all the date pickers in `$editFields` was unnecesssary,
and actually was misleading because the stewardship date picker is not in `$editFields`.

So now we're only calling `datepicker()` after details are refreshed, and we're calling
it on all date pickers (not just those that are also in `$editFields`). Sheesh.