Message makes tree diameter too narrow to use
@lederer recommended we display the message beneath the
field label because there is plenty of room there.

Also, the message should be in red, like the message
in the add-a-tree sidebar wizard.

Also, the label should not be vertical-align middle.
Its baseline should be even with the baseline of the
text in the second table cell, whether it's in edit
mode or not.

This needed to be done for all rows in the detail pages.
Otherwise, the presence of a message on the right
on any row would make the tree diameter too narrow to use.

Changes:

opentreemap/treemap/js/src/lib/
* editableForm.js
* inlineEditForm.js
* simpleEditForm.js
------------------------------

Manipulate a [data-mode] element to be either edit or display.

opentreemap/treemap/templates/treemap/partials/
map_feature_detail_base.html
----------------------------

Set #map-feature-form to have a data-mode attribute
* plot_detail.html
* resource_detail.html
----------------------------

Add a can-inline-edit class to the tables to distinguish them
from ordinary tables for styling purposes.
Specifically, it distinguishes them from the
diameter-circumference control, which uses table layout.

opentreemap/treemap/templates/treemap/field/
* attrs.html
* tr.html
* diameter_tr.html
* species_tr.html
-----------------

Add a css_class variable to attrs.html to use in the class attribute,
as distinct from the class variable already in use for the
data-class attribute.

The *tr.html templates set css_class to `alert alert-danger`
on the error elements to make them styled with red.

assets/css/sass/partials/pages/_treedetails.scss
------------------------------------------------

Remove the diameter-row style.
Style the first td in a row differently depending on whether
* the table is in edit mode
* the table contents can be edited inline
* the row the td is in is itself editable

Attention to whether the table contents can be edited
allows the style to apply just to the desired table's cells,
not to cells in child tables such as the
diameter-circumference editor.

--

Connects to #2754