Tree detail error message vs. diameter field
When there was a tree detail error, it was displayed in a new third
table column.
This caused the diameter/circumference edit widget to get so narrow you
couldn't see the values.

Changes:
1. Show the message in a new <tr>.
2. Show/hide the <tr>.
3. Show the message in a colspan="2" <td> in the <tr>,
   so it fills the width of the row.
4. Show the message in a div inside the <td colspan="2">
   to get the styling right.
5. Style it.

The styling problem is that bootstrap.css orders
.table > tbody > tr > td later than .alert-danger,
resulting in a td-looking top border, instead of
4 alert-danger borders, which looks awful.

The alternative would have been to specify the border-top
in the td[colspan="2"] style,
but then we have to copy-paste from bootstrap,
which has maintainability problems.