Combine add tree and add resource buttons into a single button group
Frees up room on the stats bar in medium sized screens.
To add a resource, you now must select it from a dropdown menu accessible
from the add tree button group, which will preselect that resource type.

Most of the changes to the javascript code are to support a mode "type"
in addition to a mode "name".  We can probably simplify this a bit by
eliminating the begining "select type" step for resources in a later
refactor.

Connects to #2737