Fix advanced search category highlighting
The "Tree" category wasn't being highlighted because the expression `_.map(search.display, R.toLower)` should have said `R.toLowerCase`, however after making that change it still failed so I wrote it using the native JS `toLowerCase`.

The "General" category wasn't being highlighted because the "Last Updated" filter has a `featureName` of `mapFeature` which wasn't accounted for in the comparison to display filters. (There is no display filter for `maspFeature` since everything is a `mapFeature`). I added a case for that.

Connects #2433