Header login link preserves current page
The header login link href now specifies the login page,
with the urlencoded current page in the `?next=` query string param.

Also changed the `InstanceUser` `__unicode__` method to do something
sensible when there is no logged in user.
Needed that for the django debug toolbar templates tab,
which was crashing in `InstanceUser.__unicode__`.

--

Connects to #2239