Fix intermittently-failing test
We were calling `instances_closest_to_point` with four instances, adding another instance, and calling `instances_closest_to_point` again. The second call was intermittently failing because the distance is 0 to two of the instances, making for an indeterminate ordering when sorting by distance. (The distance is computed to the instance bounds, and it's 0 for two instances because the point is inside their bounds.)

I removed the second call because it does not appear to test any functionality not covered
by the first call. The results of the first call don't have the indeterminate order problem.

Connects #2698