Fix tests which rely on order of collection field values
The values in a collection field (like Stewardship) are unordered. That was a surprise to me, but we get away with it because the collection fields we've used (Stewardship, Alerts) always have a date member and we can sort by that.

When we reference the values of a collection field, like here:
https://github.com/OpenTreeMap/otm-core/blob/develop/opentreemap/treemap/tests/test_udfs.py#L1087
their order is indeterminate. Here's where we build the list of values:
https://github.com/OpenTreeMap/otm-core/blob/develop/opentreemap/treemap/udf.py#L935

On the other hand, note that the order of `plot.audits()` is deterministic, since we can sort by `created`:
https://github.com/OpenTreeMap/otm-core/blob/develop/opentreemap/treemap/udf.py#L1117

To fix the tests:
* Use sets when comparing values of a collection field.
* Use a filter to remove a particular collection field value.

Connects #2700