Do not mutate the static SPECIES list
The `SPECIES` list defined in opentreemap/treemap/__init__.py is
intended to be a constant. The `add_species_to_instance` function was
mutating it, adding full `Instance` objects to each species
object. This came to our attention when some other code was trying to
call `json.dumps` on the species dict and the `Instance` objects are
not JSON serializable.