Fix quick edit move tree
Check whether the plot is writable, not specific field.
The `geom` field is now always writable, and so gives no
indication of whether the user has permission to move it.

This fix will result in navigation to the login page
for the anonymous user, and fail for a user who has
no authorizable field permissions on the plot.