Fix comment submit date handling in migrator
Threaded comments were being migrated incorrectly. The `submit_date`
was being set to the date of the migration, not the original comment
submit date. The solution is to ensure that the renaming of
`date_submitted` (which is the name used in the OTM1 fixtures) to
`submit_date` is configured in the standard migration rules for both
comments and threadedcomments, and the the mapping old name to new
name is correctly defined.