Improve performance of instance and user edits pages
The edits page had exceptionally poor performance when paging beyond the first
few pages, to the point that the requests would time out when attempting to
reach pages past the first dozen or so.

This is primarily caused by the complex query we use to ensure that users only
see edits for fields they have permission to view, by was greatly exacerbated
by the use of `OFFSET` to get pages beyond the first, which requires the
database to retrieve and discard all rows for previous pages.

By relying on the fact that we show edits in a deterministic order (newest to
oldest), we can switch from passing a page parameter to passing the starting
`Audit.id` and eliminate the `OFFSET`, making performance on each page roughly
equal.

One downside of this is that we need to keep track of the previous page
starting positions for every previous page, which means the URL can grow
to be rather long, but this seems like an OK trade-off to me.

Connects to #1818