Search by user
Changes:
- Add a `verbose_name` to `updated_by`
- Add `updated_by` to default search fields
- Add foreign key handling to `form_extras` `SearchNode`
- Add HTML to express `SearchNode` foreign key fields
- `searchBar.js` translation of foreign key DOM into `otmTypeahead` options
- `otmTypeahead` now handles dynamic search without prefetch
- `search.js` adapts foreign key fields to tiler requests
- Reset resets both significant parts of the typeahead DOM
- Style fixes
- Remove a no-op bacon onValue

Complaints:

The General dropdown doesn't expand to include the typeahead
drop-down menu, which now appears to overlap the bottom of it.
Not pretty, but usable.

Use of the `.no-hamburger` css class is not elegant.

Advanced Search fields that cannot be disabled in Management
should probably be hard-coded, not generated.

--

Connects to #388