Prevent active search icon from overlapping search box
I shortened the location search box by a few pixels and added a margin
to the search toggle, which should prevent this.

This will unfortunately clip the location search placeholder text on
skinnier phones like the iPhone 5 by a few pixels, but I couldn't think
of a good way to avoid that.

Connects to #2947