Restore instance API field identifiers
Versions of the iOS application use the presence of identifiers in the
`fields` dictionary returned from the instance info API to determine
whether or not a field in the search configuration should be invisible
to the currently logged in user. The identifiers in this commit were
not included in fields dictionary for instances created after the
transition to model-level permissions, but the identifiers are
referenced in the default mobile search configuration. Species and
photos are always visible, so we can hard code a non-None value for
these identifiers to restore the structure assumed by the iOS
application.