Strip query strings from tree photo URLs
The query strings present on S3 URLs allow for temporary access to private
buckets, but this isn't required for tree photos, which are stored in
a public S3 bucket.

For some reason having the query string present was preventing recent
versions of the Android app from downloading photos. In order to support
these older versions we're stripping the query string from the URL in the
backend, as the Android app can download photos without the query string.

NOTE there is supposed to be support in Boto for turning off signed URLs,
but this appears to be broken when using IAM roles instead of an access key
and secret key for authentication with AWS.

Connects to OpenTreeMap/otm-android#307