Attempt to fix intermittent test failure
The failure suggests that when typing "32" for diameter and hitting "Save", the new diameter is sometimes not saved. My only thought is that the "save" event is handled before the diameter calculator's JS code has finished updating everything. So, wait for the circumference value to be displayed before continuing.

Also fixed an unrelated oversight in `browse_to_instance_url`.

Connects #2982