Fix polygon area calculation
The `getPolygonDisplayArea` function began returning zero when passed
the results of calling `getLatLngs` on a boundary polygon. The
`getLatLngs` method was producing a doubly-nested array and the
`geodesicArea` function only works on a 1-dimensional array. I fixed
`getPolygonDisplayArea` by making it recursively descend through an
arbitrary number of nested arrays and sum the results of calling
`geodesicArea` on any "leaf" array it finds.