Make sequence number fields on Instance readonly
Use F expression to prevent Django from overwriting sequence numbers
in the database with old values. The only way to increment these values
should be to use the correct helper methods.

Note that we have to take care to NOT use F expressions on records that
have not been inserted into the database yet. Trying to do so results in
an exception which prevents Django from inserting the object.