Stewardship with date search fix
Searches that include multiple queries on a single collection
UserDefinedFieldDefinition need to be grouped together into
a single subquery.

The previous search created a separate subquery for each one,
and so incorrectly matched entities that had one
UserDefinedCollectionValue that matched the action,
another that matched the date, but none that matched both.

This fix necessitated test changes.

--

Connects to #3029