Force CSRF cookie to be sent on Map Page
Normally this cookie would already be present from other actions / pages,
which have forms that use the `{% csrf_token %}` template tag, but if the
Map page is the first page visited (as when using an Icognito window) it would not be set.

This wasn't noticed before because until polygonal search was added, there were no actions on the map page that required a CSRF token but did not first require login (which would set the cookie).

Adding this decorator to the route forces the cookie to be set, fixing the issue.

Connects to #3093