upgrade HStore
Overview
--------

The point of this commit is to switch from
`django_hstore.fields.DictionaryField` to
`django.contrib.postgres.fields.HStoreField`.

The former is incompatible with later releases of django,
and OpenTreeMap needs to migrate from the current 1.8 to 1.11
for stability.

The latter is the sanctioned django interface to PostgreSQL's hstore
going forward.

The latter is also less functional than the former, prompting a series
of other changes.
Changes were also made for the sake of maintenance simplicity.

The key changes here are in `treemap.udf`, `treemap.audit`, `*.models`, and
`treemap.search`.

**treemap.udf**

Entry points:
-   `UDFModel`, a base class for models that need user defined fields
-   `UserDefinedFieldDefinition`, defines the name, data type, and model type
    for a user defined field, for a given treemap instance, as a
    scalar or a collection
-   `UserDefinedCollectionValue`, an HStore collection value for a
    specific `UserDefinedFieldDefinition` on a specific model instance

`UDFModel` user guide:

Classes that want custom fields subclass `UDFModel`.

To assign a custom field,
    `my_model_instance.udfs['custom field name'] = value`

There are three ways to retrieve it:
    `my_model_instance.udfs['custom field name']`
    `getattr(my_model_instance, 'custom field name')`
    `getattr(my_model_instance, 'udf:custom field name'`)

All of the above work for both scalar and collection custom fields.

Filter `UDFModel` subclass querysets using the `udf` transform.

In addition to the transforms and lookups described at
https://docs.djangoproject.com/en/1.8/ref/contrib/postgres/fields/,
`UDFModel` implements `__int` and `__float` transforms, for use
before magnitude comparisons (`__gt`, `__gte`, `__lt`, `__lte`).
Without them, comparison is lexical, which does not yield the
same results as a numerical comparison.

For example, to filter a queryset for a custom field value
corresponding to an integer `UserDefinedFieldDefinition`
that should be less than some ceiling, do

```
MyUDFModelSubclass.objects.filter(**{
    'udfs__' + custom_field_name + '__int__lt': ceiling})
```

**treemap.audit and \*.models**

- Models can use a plain `GeoManager` now
- In order to reduce code coupling and assure that initialization happens in a
  rational order, `populate_previous_state` is now the responsibility of the
  leaf subclass of `UserTrackable`.
- Species is no longer a `UDFModel` because it never employed udfs,
  and removing unused code makes maintenance simpler.

Before this change, `populate_previous_state` was getting called between
initialization of different levels of model class. It calls `as_dict`, which
raises exceptions if not all fields have been initialized,
so now `populate_previous_state` is the responsibility of the leaf subclass.

This reduces the coupling between auditing and custom fields.

That tight coupling was why previous versions of `udfs.py` had to
inject a descriptor into the django model instantiation procedure.
Descriptors are fine as long as they work, but when they don't, they
are a headache to debug. Better to reduce the amount of injection
into django's procedure.

**treemap.search**

- Use the `udf` filter transform
- Additional parser steps to add the udf `__float` filter transform

--

Connects to #1968