Replace django-url-tools dependency with new app code
The django-url-tools package is unmaintained. We have only made use of
a single template tag from this package so instead of forking the
library, building out the test infrastructure, and making enhancements
to ensure compatibility with Django 1.11, I have opted to bring the
single template tag and supporting code into the application.