Skip tree creation if tree UDF values are empty
We previously added code to protect against the iOS app always sending
values for all UDF fields. While this protective code does skip over
the values, it only runs after the presence of a `tree:` field in the
request triggers the instantiation of a `Tree`, which then gets saved.

In this commit I have added a check to skip field value processing
entirely and bypass the creation of a new `Tree` in the situation
where the `Tree` does not already exist and the field value is empty.