Rewrite migrations to remove dependency on django_hstore
This is a bit hacky, but it seems to work.

We've switched to using the built-in HStore support from Django, so the only
reference left to django_hstore is in our migrations

The root of the issue is that django_hstore doesn't work with Django 1.11,
but it was still referenced from the initial migration in treemap.

Since no actual changes happen at the database level when switching from one
HStore field to the other, I've edited the initial migrations so that the
column in question was "always" using the built-in HStore field