Add link from modeling scenarios to prioritization demo
We extracted the modeling prioritization functionality into a separate demo
application. In this commit we link to that application, passing the instance
center coordinate so that prioritization starts in the same location as the
instance.

After some design iteration we decided to move the scenario burger button to the
far left of the toolbar to make room for the new prioritization link.