Update UDFDictionary so that setting key to None does not raise
The authorizable system sets field values to none if the logged in user is not
permitted to see the field. In the case of UDFs, setting a field value to None
when the key did not exist in the UDF dictionary was raising a KeyError. This
commit avoids the error by checking for the existence of the key before trying
to remove it.