Update otmTypeahead track last menu selection / skip autocomplete
In a previous commit we made the search system more user friendly by
automatically triggering an autocomplete when pressing the enter key in a
typeahead or clicking the "Search" button. In the case where the user selects an
item from the dropdown list rather than typing Enter, clicking "Search" would
trigger the autocomplete action which would select the first item in the
dropdown list instead of acknowledging the already selected item.

This commit addresses that problem by keeping track of the last selected item
and skipping the autocomplete action if the current value of the text field
matches.