Add modeling permission data migration
Any permissions defined in code must have a supporting `Permission` row.

This migration is unusual because it contains logic that was originally created
in a different Django application. As a result, in some environments the change
has already been applied. To deal with this situation, I added a try/except
block to exit the migration early if the `Permission` row already exists.