Return plot create/update validation errors with a 400 status
Previously, validation errors in the plot create and update views were not being
caught, resulting in a 500 error being logged in Rollbar.

Updating these views to properly handle the validation errors required a fix to
the `transform_plot_update_dict` function to check for the existence of a key
before trying to read it from the plot dict.