Store geocoded address in URL when searching and show on page load
Adds a new `a` parameter to the URL which is set when there is a
non-boundary value in the location typeahead.

On page load on the map page, if the `a` parameter is present in the URL
and no boundary filter is present, we geocode the address and display a map
pin (which will also recenter the map on the pin).

Connects to #1123