Fix streetview update after saving on the tree detail page
There was existing code to update streeview after a successful save, but that
code expected the form data to include `plot.geom`. It does not. `plot.geom` is
appended directly to the data PUT back to the server in a before save handler.

To get around this, we just read the location from the plot mover. It is safe to
call `panorama.update` even if there is no geometry change. It is effectively a
noop.