Don't show the tree photo sharing dialog if it is not appropriate
There are three tree photo upload situations where we do not want to show a
social media sharing dialog after POSTing the image to the server.

- The instance is private.
- The image file was invalid.
- The image file was too big.

Private instances are not handled by not rendering the dialog markup on the page
unless the instance is public.

Invalid images are handled by checking the markup returned from the response for
the presence of an error element. This is not ideal, but is easier that changing
the way failures are reported vs. loading updated carousel markup.

Large images are intended to be blocked by Nginx before reaching Django. If they
reach Django they will trigger a validation error, which will be handled the
same way invalid images are handled.