Add filter unit conversion
Prior to this commit filters did not take instance unit preferences into
account. I have added a "preprocessing" method that we can pass the filter
object through which converts any values before the filter object is passed off
to be converted into a query.

I have decided to implement this conversion on the server side to prevent having
to change both mobile apps and the web client.