Update universal rev after changing instance units.
Calculating ecobenefits for a large number of trees can be expensive, so we
cache the results based on query string arguments and the `universal_rev` of the
instance. After changing the units of a field, the same query string can produce
different results so we must increment the `universal_rev` to bust the cache.