Update jquery.fileupload to 9.12.4
The old version was triggering deprecation warnings after installing
jquery-migrate.

I tried upgrading to the latest version as of this commit, 9.20.0, but it caused
a webpack error due to a change in AMD require naming.

https://github.com/blueimp/jQuery-File-Upload/commit/a9bfde1cfe7c5295b3757a28fbffe06087701749

I pulled in the version before this naming change (9.12.4) and it resolved the
webpack issue along with eliminating the deprecation warning from
jquery-migrate.

I also updated the jquery.iframe-transport.js and jquery.ui.widget.js
dependencies with versions from the 9.12.4 distribution.