Replace deprecated $.bind usage with $.on
Cause:: These event binding methods have been deprecated in favor of the .on()
and .off() methods which can handle both delegated and direct event binding.
Although the older methods are still present in jQuery 3.0, they may be removed
as early as the next major-version update.

Solution: Change the method call to use .on() or .off(), the documentation for
the old methods include specific instructions. In general, the .bind() and
.unbind() methods can be renamed directly to .on() and .off() respectively since
the argument orders are identical.

https://github.com/jquery/jquery-migrate/blob/496bd8738785e6c88248487302dee7dd00607619/warnings.md#jqmigrate-jqueryfnremoveattr-no-longer-sets-boolean-properties