Monkey patch the Leaflet layer selector to work in Mobile Safari
Tapping on the layers control in Mobile Safari was not opening the layer
selector. I was not able to reproduce this issue in a separate test application,
only within OTM.

After much trial and error I discovered that changing this line
  L.DomEvent.on(el, L.Draggable.START.join(' '), stop);
to
  L.DomEvent.on(container, 'mousedown', stop); in
DomEvent.disableClickPropagation resolved the issue. Effectively, this
prevents the `touchstart` event from being stopped.

I opted to implement this as a runtime patch applied within MapManager instead
of a global patch to the Leaflet source because there are no other Leaflet
behaviors that are known to be broken in Mobile Safari and I did not want to
risk introducing a regression in some other event handling.

I minimized the chance of forgetting about this patch if/when we upgrade our
version of Leaflet by adding a comment to package.json.