Allow a hidden Tree ID field to be shown
The management page was allowing the Tree ID field to be hidden, but was not
rendering the option in the hidden field list so that it could be shown again.
We have corrected the `_should_show_field` predicate function and corrected the
model argument passed to it so that the Tree ID field is correctly shown in the
disabled field list.