Fix can_edit_tree instance meta permission
The Android application checks the value of the `can_edit_tree` instance meta
permission and does not PUT changes to the API if it has a value of false. This
meta permission was incorrectly looking only at plot fields. This resulted in
members of a role that could only edit tree fields being prevented from saving
any tree changes.