Fix variable references in list comprehension for Python 3
List comprehensions in Python 3 create a new functional scope. As a result in
the case where we are using a list comprehension when defining a class-level
member we do not have access to the other class-level members within the
comprehension. We have implemented a workaround that uses an immediately
executed lambda to bind the member variables in the new scope.