Avoid DotDict exceptions due to Python 3 change in hasattr handling
Django 1.11.17 calls `hasattr(a_dotdict_field, 'resolve_expression')`.
https://github.com/django/django/blob/121115d2c291b3969ac00ca62253f23513481739/django/db/models/sql/compiler.py#L990

The Python 3 documentation says that this ends up calling `getattr` and
returning False if an `AttributeError` is raised
https://docs.python.org/3/library/functions.html#hasattr

For dictionaries `getattr` ends up calling `__getitem__`. Our implementation of
`__getitem__` correctly raises KeyError, and we were relying on a bug in Python
2 where any exception raised during a `hasattr` check would result in False. Now
that Python 3 explicitly checks for `AttributeError` we need this workaround

The tradeoff is that we now return `AttributeError` instead of `KeyError` when
attempting to access a non-existent key. If we rely on catching `KeyError`
somewhere in our application we can fix it there.